Upgrade dataloader

+++ src/main/java/com/aclab/dne/model/MetaMatiere.java
@@ -19,19 +19,19 @@ public class MetaMatiere  implements Serializable {
 
     @Id
     @Column(name = "id_promotion")
-    private Long idPromotion;
+    private long idPromotion;
 
     @Id
     @Column(name = "id_diplome")
-    private Long idDiplome;
+    private int idDiplome;
 
     @Id
     @Column(name = "id_ue")
-    private Long idUe;
+    private int idUe;
 
     @Id
     @Column(name = "id_matiere")
-    private Long idMatiere;
+    private int idMatiere;
 
     @ManyToMany(mappedBy = "matieres")
     private Set<Enseignant> enseignants;
